tweaks following feedback
